♻️ Replaced `PopoverVariant`
