+++ packages/react/src/components/Popover/index.ts
@@ -1,2 +1,2 @@
 export type { PopoverProps } from './Popover';
-export { Popover, PopoverVariant } from './Popover';
+export { Popover } from './Popover';
